requirements.txt
django==3.2.12
psycopg2-binary==2.9.*
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.1
pip==22.1
setuptools==62.3.2